Use single Input component
