--- src/app/components/InputWithLabel.tsx
@@ -1,41 +1,0 @@
-import React from 'react';
-
-interface Props extends React.InputHTMLAttributes<HTMLInputElement> {
-  containerClassName?: string;
-  inputClassName?: string;
-  label?: string;
-  id?: string;
-  loading?: boolean;
-}
-
-export function InputWithLabel(props: Props) {
-  const {
-    containerClassName = '',
-    inputClassName = '',
-    loading = false,
-    label,
-    id,
-    ...restProps
-  } = props;
-  return (
-    <div className={`mb-2  ${containerClassName}`}>
-      {label && (
-        <label
-          className="block text-gray-700 text-sm font-bold mb-2"
-          htmlFor={id}
-        >
-          {label}
-        </label>
-      )}
-      <input
-        className={`shadow appearance-none border rounded w-full py-2 px-3 text-gray-700 leading-tight focus:outline-none focus:shadow-outline ${inputClassName}`}
-        id={id}
-        type="text"
-        disabled={loading}
-        {...restProps}
-      />
-    </div>
-  );
-}
-
-export default InputWithLabel;
